pyproject.toml
[tool.poetry]
name = "spamoverflow-27abdullah"
version = "0.1.0"
description = ""
authors = ["27abdullah <abdullahbadat27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.2"
flask-sqlalchemy = "^3.1.1"
psycopg2 = "^2.9.9"
find_domains = "^0.0.6"
rfc3339 = "^6.2"
iso8601 = "^2.1.0"
gunicorn = "*"
celery = {extras = ["sqs"], version = "^5.3.6"}
redis = "^5.0.3"
boto3 = "^1.34.85"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"